refactor(team): move query accessors into entity files
